requirements.txt
attrdict==2.0.0
bleach==2.1.2
certifi==2016.2.28
click==6.7
contextlib2==0.5.5
cycler==0.10.0
decorator==4.2.1
docopt==0.6.2
entrypoints==0.2.3
enum34==1.1.6
Flask==0.12
future==0.16.0
git-review==1.25.0
gitdb==0.6.4
GitPython==2.0.8
html5lib==1.0.1
humanize==0.5.1
imgaug==0.2.5
ipykernel==4.7.0
ipython==6.2.1
ipython-genutils==0.2.0
ipywidgets==7.1.0
itsdangerous==0.24
jedi==0.11.1
Jinja2==2.10
jsonschema==2.6.0
jupyter==1.0.0
jupyter-client==5.2.1
jupyter-console==5.2.0
jupyter-core==4.4.0
kitchen==1.2.4
MarkupSafe==1.0
matplotlib==2.1.1
mistune==0.8.3
mock==2.0.0
more-itertools==2.2
names==0.3.0
nbconvert==5.3.1
nbformat==4.4.0
neptune-cli
networkx==2.0
notebook==5.3.1
numpy==1.14.0
oauthlib==2.0.6
opencv-python==3.4.0.12
pandas==0.22.0
pandocfilters==1.4.2
parso==0.1.1
pathlib2==2.3.0
pbr==3.1.1
pexpect==4.3.1
pickleshare==0.7.4
Pillow==5.0.0
prompt-toolkit==1.0.15
psutil==4.3.0
ptyprocess==0.5.2
pydot==1.2.4
pydot-ng==1.0.0
Pygments==2.2.0
PyJWT==1.5.2
pykwalify==1.5.2
pyparsing==2.2.0
python-dateutil==2.4.2
pytz==2017.3
PyWavelets==0.5.2
PyYAML==3.11
pyzmq==16.0.3
qtconsole==4.3.1
raven==6.1.0
requests==2.11.1
requests-oauthlib==0.8.0
scikit-image==0.13.1
scikit-learn==0.19.1
scipy==1.0.0
Send2Trash==1.4.2
simplegeneric==0.8.1
six==1.10.0
smmap==0.9.0
terminado==0.8.1
terminaltables==2.1.0
testpath==0.3.1
torchvision==0.2.0
tornado==4.2
tqdm==4.11.2
traitlets==4.3.2
urllib3==1.16
voluptuous==0.9.3
wcwidth==0.1.7
webencodings==0.5.1
websocket-client==0.35.0
Werkzeug==0.14.1
widgetsnbextension==3.1.0
